Simplify tox dependencies (#5)
This does not address the issues in an expected way. The issue recommends that we update the documentation about how to add dependencies to the tox.ini file. However, I felt that this still requires the user to maintain multiple dependency lists in multiple locations, thus not simplifying the issue. I have utilised a tox.ini feature to move the dependency list into the setup.py file.

I also believe that the setup.py repository should be understood by general python programmers and thus we do not need to explain this change in the checklist.